Added server.cfg. Improved Server.shutdown()
Server is now configurable via server.cfg. The Config object is a
singleton which provides the validated configuration data. It is
initialized via its init() method.

I also added a ThreadGroup to Server. The constructors of WelcomeThread
and ClientThread use this to place themselves into the Server's
ThreadGroup. This vastly simplifies Server.shutdown().