Server's config file has options for clients
Config file has user_ids and private_keys options, which are
comma-separated lists of the ids and private keys of clients.

They use the Server.initDB() function. Because of this,
the Server class now requires the config file to be passed into it.
Some code changes were made to Server as a result of this.