Implemented RESPONSE parsing in server.
The RESPONSE is parsed, but not yet cryptographically validated.  In the
process, all of the protocol messages have been "stringly typed". When
arbitrary binary such as hashes are sent/received, they are converted
to/from hex strings. This makes the parsing much easier because
everything can be done with Scanner and DatatypeConverter.