Implemented encryption.
For complicated reasons, I had to change the underlying algorithms for
hash1 and hash2 to make this work. Hash1 is now SHA-1 and hash2 is
MD5. Encryption is AES (128-bit). I would have preferred AES-256, but it
is not implemented in standard Java, and libraries for it seem to be
impossible to come by.