Added client_addr/port to client.cfg
In preparation for implementing REGISTER:

The client needs to explicitly advertise an IP address and port that the
server should "call back" to. This is problematic since the client
software can't know its external IP address if it is behind NAT.

To make a long story short, this necessitates some new settings in the
client: client_ip, client_external_ip, and client_port.