Cleanup client HandshakeThread. Bugfix.
Split up the run() method into smaller parts paralleling the changes
made in server HandshakeThread in the previous iteration.

Also realized a potential race condition between the HandhsakeThread
and SessionThread in both the client and server. It is now guaranteed
that the HandshakeThread cannot try to _client.setState() after the
SessionThread has been spun.