Fixing one more race condition.
It was barely possible for the client to send REGISTER from its
HandshakeThread, and for the server respond back with REGISTERED before
the SessionThread on the client managed to get the port fully opened and
in the accept() state.

There's not a perfect way to deal with this since the Java Socket object
itself does not have an isListening() function (just isBound() and
isClosed(). I've fixed this with a semaphore that blocks the client
HandshakeThread from sending REGISTER until just before its
SessionThread calls into Socket.accept().

This only seemed to happen when you launched two clients and logged in
from both in very rapid succession over loopback (localhost) but
even so...